Big refactor for driver averaging
New helper Accumulator class to handle the data with int64 datatype
snapshots up to the top.

--CAR and BRJ

+++ hardware/X6-1000M/src/X6ADC/libx6adc.h
@@ -60,7 +60,7 @@ EXPORT int set_averager_settings(int, int, int, int, int);
 EXPORT int acquire(int);
 EXPORT int wait_for_acquisition(int, int);
 EXPORT int stop(int);
-EXPORT int transfer_waveform(int, int, short *, unsigned int);
+EXPORT int transfer_waveform(int, int, int64_t *, unsigned int);
 
 EXPORT int set_log(char *);
 int update_log(FILE * pFile);
